Reduce cost of Bandages, Burn and Trauma Kits
This should reduce the cost of the afforementioned items. It'll also be my first code contribution to a Space Station 13 server.
The reduction is as follows:
Bandages: From 2 cloth to 1 cloth.
Burn Kits: From 4 cloth to 1 cloth.
Trauma Kits: From 5 cloth to 2 cloth.
This is to hopefully allow for  easier access to advanced medicine without having to harvest great swaths of cotton to carry the cost.
Values suggested by Admiral Taislin and Sgt. Shinobi on the Civ13 Discord.

Did I do this right?