Adds some of LanceLynxx's Sounds
This should add some of the sounds LanceLynxx requested be added on the Discord, including wolf, bear, and panther sounds, snow and ice sounds, and pickaxe sounds among others. There are other files to add in to the game, but I'm on a strict 'dozen lines per fourty-eight hours' coding diet, so I'm afraid I'll have to schedule those away for later.