Fixes/ Added more stuff
Removed Gaelic spawns from Nomads_Europe so the pop is not as divided (Island is now neutral), added prehistoric creatures to Nomad_ice_age and I fixed and added a bunch of badges. Adds Food/plant sprites that havi made and I added the other Shotgun shells that Alpha made.