Updates the Gigantopithecus sprite and switches the map to load updated colony
Moved Colonie's carib spawn to a new location and gives them premade buildings for a start.

Adds Bison to the map and removes a few mice spawns.

Bisons are planned to just be the main food of the tribals till the colonists hunt them to death since they have no spawns and should breed.