Adds edits to maps and sprites.
Removes the extra walls from Colony/Jungle Colony.

Replaces the Ocean Area of Colony and Jungle Colony with the proper weather area so you can't just walk into the ocean and avoid the blizzard.

Minor edits to NationsRP, removes bear spawns from the middle of the river and a few tree spawns that where in the river not on land, Minor edits to teutoburg to allow the Barbarians to build a bridge to fight the romans.

Fixes the current stone hatchet and stone pickaxe so they look more like stone.

Might plan a flint pickaxe or something.