Stack Fixes
A stack that is created in the attack_hand() proc
will now use the amount of the older stack for the transferral if this
amount happens to be less than one.

A stack is no longer immediately deleted when you try to use it to craft
something while it has an amount that is between zero and one.