Adds Lard Soup + Soup routines for players to bath with it
Lard soup recipe will come after the Lye PR takes effect.

A soap will get the hygiene to the maximum(250) and boost your mood(same as eating 5 dried meat).
Theres a 15 minutes  cooldown between each soap mood, so people wont abuse it.
You need to be in any water tile to take a bath with it.

It holds 5 sprites that shows up depending on how used the soap is. It will be totally used, and disappear after being used 30 times.