Copypaste of fur_coat code
Copypasting fur_coat code for monk_robes to make hair display normally on hood switching. Maybe it would be better to not create excess else if blocks for same piece of clothing, and just add some ORs to fur coat lines, but i'm not sure. It is what it is