Extends the Little Creek RP map
Adds a canyon to the south of the map, and a farm house.

Moved the Graveyard right next to the Church.

Moved the Cattle field to the South with the new Farm house.

adds a longer road to the south and an old mineshaft tunnel.

Adds a WIP Indian settlement for a Indian Faction that will be togglable.

Pretty much four buildings with items within it, no indians to live in it though.