Fixes Grammatical issues and adds a new turf/ edits the map of the tantive to be more appealing.
Replaces the Blast doors with Generic walls so they Stormtroopers can actually get through with their explosive charges.

Fixes some lines of code to make it fit in more with the setting and fixes the spelling of Imperials from "impierals"

Adds a space turf, and makes the windows able to be seen out of instead of a void.

Should be 85% there and playable.