Adds more to the map.
Fleshes out the map more.

Adds repeating mounted laser weapons in two points of the map that'll help the rebels a bit.

Added more cover for the rebels.

Added a few secret easter eggs on the map as gear.

Adds Blast doors north to ease the Imperial Push.

Fixes the on mob sprite for the E-11 a bit.