Fixes some Things with VantiveIV
Fixes a few mess ups in the map_metadata so it reads a bit better so its not "Imperial Shocktroopers Troopers"

Adds StormSurgeon Armor/ Changes Imperial Doctor role to StormSurgeon

Adds a Mint Green StormSurgeon Armor, Basically ShockTrooper Armor with green accents, its a bit weaker then basic armor yet a bit more fast to move in.

Makes it so Barricaded Windows on VantiveIV is pretty much invincible so players can't break through them.

And edits a few lines in the job file to match the setting a bit.