WIP Elevators (working, need time adjustments and maybe exploit fixes)
Acts as a teleporter for anything (except the elevator buttons and obj/covers/ types in the range of tile 1 of the interior button. If the elevator is 2x2,  place the interior button in the upper left or right corner, use pixel_y and pixel_x to place it against the wall.
If the elevator is 3x3 (no such doors sprited yet), place it in the middle and do pixel_y = +64 to have it on the wall.