WIP beekeeping
Everything works, except the process() proc, which is responsible for producing honey... But that's a problem for all objects having a process (i.e. ported defib, candles, etc.).
Gonna dig in Bier's new "additions" to the processing code on a later day. If I don't find a solution, a shitty way would be to make an endless loop called on New()...