Temporary Nomads North America spawns
Caribs spawn on southern parts of the map, Cherokee (which is false, should be in the South East of the USA area) spawn on the western parts of the map, Iroquois spawn on the Eastern parts of the map (should be North East of the USA area).

At the Imperial Age and so on, Spaniards/Mexicans spawn in the Southern part of the map, English speakers spawn on Northern part of the map. Probabilities to spawn as a Native American are reduced.

To-do: Inuit spawns in the Northern parts of the map, Hawaiians in Hawaii, add more Native American cultures like Apaches, Cheyenne, Sioux, maybe add the French in the North West.