Hotel Revamp
This pr changes Hotel alot, changes include but are not limited to.

Hotel remap

Minor decorating in surrounding areas

Attempting to improve the overall feel of Hotel by removing the holes in the objective and adding new points of entry aswell as making current points of entry more even between the two sides.